requirements.txt
asgiref==3.2.10
Django==3.1
django-autofixture==0.12.1
django-bootstrap3==14.1.0
django-formtools==2.2
djangorestframework==3.11.1
factory-boy==2.12.0
Faker==4.1.1
mysql==0.0.2
mysqlclient==2.0.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
South==1.0.2
sqlparse==0.3.1
text-unidecode==1.3
